Reset the code base to use the current master state in vibe.d.
Instead of trying to use the new code base that evolved from the HTTP/2 implementation, the plan is to first separate the existing code out to the vibe-http package, where regular development can continue. This state will be tagged as 1.0.0. At the same time, a different branch will be used to integrate the HTTP/2 code, which will at some point lead to a new 2.0.0 version tag.