Fix handling of HTTP/1 requests with both, Content-Length and Transfer-Encoding.
In order to avoid the possibility of request smuggling or response splitting and in accordance with RFC 9112 6.3.3, the "Transfer-Encoding: chunked" header must have precedence over any "Content-Length" header that is also present. The if-else chain that handles the precedence has been reordered accordingly. Additionally, this also makes use of the possibility to reject such a request altogether and replies with a 400 status. Since vibe.d is usually used in conjunction with a controlled HTTP client, such as a reverse proxy, the vibe.d HTTP client, or a common web browser, the likelihood of this causing breakage should be low.

See https://github.com/vibe-d/vibe.d/security/advisories/GHSA-hm69-r6ch-92wx